% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{add}
\alias{add}
\alias{minus}
\title{Functions for adding and subtracting}
\usage{
add(x, y)

minus(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
a number
}
\description{
A really good adding/subtracting function.
If Carlsberg made adding functions, they'd make a function like this.

A work of pure genius.
}
\examples{
add(5, 10)
## Can also use negative numbers
add(-5, 10)
}
